Upgrades

+++ FakeLocation.Shared/UIModels/TagReadModel.cs
@@ -3,9 +3,9 @@
     public class TagReadModel
     {
         public int Id { get; set; }
-        public int X { get; set; }
-        public int Y { get; set; }
-        public int Z { get; set; }
+        public double X { get; set; }
+        public double Y { get; set; }
+        public double Z { get; set; }
         public int SignalFrequency { get; set; }
     }
 }
